Pause Menu & Input & Shipping
Added a pause menu due to capture the game on movie and needed better eprformance. But continue as it seems to be working fine. Added P & ESC to access the pause menu and Gamepad Right Special. Added intuition mode input to gamepad as well, and changed package settings to Shipping.